reborn

+++ src/entities.cpp
@@ -244,7 +244,7 @@ void checkitems() {
 			continue;
 		if (OUTBORD(e.x, e.y))
 			continue;
-		Vec3 v = { e.x, e.y, S(e.x, e.y)->floor + player1->eyeheight };
+		Vec3 v = { (float)e.x, (float)e.y, (float)(S(e.x, e.y)->floor + player1->eyeheight) };
 		vdist(dist, t, player1->o, v);
 		if (dist < (e.type == TELEPORT ? 4 : 2.5)) {
 			pickup(i, player1);
